refactor: 레이블 스타일 파일 컴포넌트 폴더로 이동
