+++ src/components/common/label/label.style.ts
